refactor: refactored parsing and started implementing known for field (#11)
