+++ utilities/Utils.ts
@@ -1,13 +1,4 @@
-import { UserModel } from "~/model/UserModel";
-
-/**
- * Given a string, this function returns the same string with the first letter capitalized.
- * @param str - the string to capitalize
- */
-export function capitalize(str: string): string {
-    const s = str.trim()
-    return s.charAt(0).toUpperCase() + s.slice(1);
-}
+import {UserModel} from "~/model/UserModel";
 
 /**
  * Gives you a random element from the list
@@ -30,6 +21,128 @@ export function getRandomUserModel(): UserModel {
     userModel.updateStats(Math.floor(Math.random() * max_value), Math.floor(Math.random() * max_value), Math.floor(Math.random() * max_value), Math.floor(Math.random() * max_value), Math.floor(Math.random() * max_value), Math.floor(Math.random() * max_value))
     return userModel;
 }
+
+// --------------------------------- String utilities  --------------------------------- //
+
+/**
+ * Given a string, this function returns the same string with the first letter capitalized.
+ * @param str - the string to capitalize
+ */
+export function capitalize(str: string): string {
+    const s = str.trim()
+    return s.charAt(0).toUpperCase() + s.slice(1);
+}
+
+/**
+ * Given a name (string) this function returns a string containing the initials of the name
+ * @param name - name we want the initials of
+ */
+export function getInitials(name: string): string {
+    const words: string[] = name.split(' ');
+    let initials: string = '';
+    for (const word of words) {
+        if (word.length > 0) {
+            initials += word[0].toUpperCase() + ". ";
+        }
+    }
+    return initials;
+}
+
+// --------------------------------- String cleaning utilities --------------------------------- //
+
+/**
+ * Remove all text  between the opening and closing tags in the given string.
+ * @param text - a string that may contain tags to remove
+ * @param opening - the opening tag
+ * @param closings - all the possible closing tag
+ */
+export function removeTag(text: string, opening: string, ...closings: string[]): string {
+    const regexStr = `${opening}[^]*?(${closings.join('|')})`;
+    const regex = new RegExp(regexStr, 'g');
+    return text.replace(regex, '');
+}
+
+/**
+ * Remove all occurrences of the given name in the given text, including
+ * first name and last name, excluding name particles (e.g. "von", "de", ...).
+ * @param name - celebrity first name(s) and last name(s)
+ * @param text - text from which to remove the name
+ */
+export function removeNameOccurrences(text: string, name: string): string {
+    // Get lists of all first names and names, without name particle (i.e. von, de, of...)
+    const names: string[] = [
+        ...name.split(" ").map(n => n.trim()).filter(n => n === capitalize(n)), name
+    ]
+    names.forEach(n => {
+        const regex = new RegExp(n, 'gi');
+        text = text.replace(regex, "???");
+    })
+    return text
+}
+
+// --------------------------------- Miscellaneous --------------------------------- //
+
+
+/**
+ * Given a sentence of words separated by "and", returns all
+ * possible permutations of the words around these "and".
+ * @param input - a sentence of words separated by "and"
+ */
+export function getAndPermutations(input: string): string[] {
+    const words = input.split(" and ").map(word => word.trim());
+
+    function generatePermutations(arr: string[], index: number) {
+        if (index === arr.length - 1) {
+            return [arr.slice()];
+        }
+
+        const result: string[][] = [];
+        for (let i = index; i < arr.length; i++) {
+            [arr[index], arr[i]] = [arr[i], arr[index]]; // Swap
+            result.push(...generatePermutations(arr, index + 1));
+            [arr[index], arr[i]] = [arr[i], arr[index]]; // Restore
+        }
+        return result;
+    }
+
+    const permutations: string[] = [];
+    const permutedArrays = generatePermutations(words, 0);
+    for (const permutedArray of permutedArrays) {
+        permutations.push(permutedArray.join(" and "));
+    }
+
+    return permutations;
+}
+
+/**
+ * Split the given text into parts containing equal number of sentences.
+ * @param text - the text to split
+ * @param num - the number of parts to split the text into
+ */
+export function splitIntoEqualSentenceParts(text: string, num: number): string[] {
+    // Split the text into sentences
+    const sentences = text.match(/[^.!?]+[.!?]+/g);
+    if (!sentences) return [text];
+
+    const totalSentences = sentences.length;
+    const sentencesPerPart = Math.floor(totalSentences / num); // Number of sentences per part
+    const parts: string[] = [];
+
+    // Distribute sentences evenly across the three parts
+    let start = 0;
+    for (let i = 0; i < num - 1; i++) {
+        const end = start + sentencesPerPart;
+        parts.push(sentences.slice(start, end).join(''));
+        start = end;
+    }
+
+    // Add the remaining sentences to the last part
+    parts.push(sentences.slice(start).join(''));
+
+    return parts;
+}
+
+// --------------------------------- Data --------------------------------- //
 
 export const months: {[key: number]: string} = {
     1: "January",
@@ -241,3 +354,4 @@ export const countries: {[key: string]: string} = {
     "Zambia": "Zambian",
     "Zimbabwe": "Zimbabwean"
 }
+
